fix: CQDG-324 family id from id service
