+++ src/test/resources/idserver/hash.json
@@ -3,6 +3,7 @@
   {"hash": "5a9855ac9875d9c1480daf476ab955abbb1f84c4", "internal_id":  "PRT0000002"},
   {"hash": "7a3532b572d958f17759970c86014593100dab02", "internal_id":  "PRT0000003"},
   {"hash": "e2a7b589db2f0d83e557e02075872e1c4a5bf271", "internal_id":  "PRT0000004"},
+  {"hash": "8cb2237d0679ca88db6464eac60da96345513964", "internal_id":  "FM00000001"},
 
   {"hash": "fd723276db594db5e56e8bc790d033f3ec561381", "internal_id":  "STU0000001"},
   {"hash": "9bc71333e51f59ff4db98ff0b1d95514c20056be", "internal_id":  "DIA0000001"},
